pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chainofcustody"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dnachisel>=3.2.16",
    "matplotlib>=3.9.0",
    "mirmap>=2.0.0",
    "mygene>=3.2.2",
    "numpy>=2.0.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pymoo>=0.6.1.6",
    "requests>=2.32.5",
    "rich-click>=1.9.7",
    "rpy2>=3.5.0",
    "viennarna>=2.7.2",
]

[project.scripts]
chainofcustody = "chainofcustody.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["chainofcustody"]

[dependency-groups]
dev = [
    "hatchling>=1.29.0",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["integration: tests that call live external APIs"]
addopts = "-m 'not integration'"
